% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.r
\name{dgp}
\alias{dgp}
\title{Generalized Poisson Distribution}
\usage{
dgp(y, mu, phi)
}
\arguments{
\item{y}{Vector of integer values, with \eqn{y \geq 0}.}

\item{mu}{The mean parameter, with \eqn{\mu > 0}.}

\item{phi}{The dispersion parameter, with \eqn{\phi \geq 0}.}
}
\value{
Vector of probability values of \eqn{Y = y}.
}
\description{
Probability function of a random variable with Generalized Poisson distribution.
}
\examples{
pg(0L:100L, mu = 0.5, phi = 0.7) |> sum()


}
\seealso{
\code{\link[=dgeom]{dgeom()}}, \code{\link[=dpois]{dpois()}} and \code{\link[=dbinom]{dbinom()}}.
}
