% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgp_pg.r
\name{fgp_pg}
\alias{fgp_pg}
\title{Probability generating function of the Geometric Poisson distribution}
\usage{
fgp_pg(t, mu, phi)
}
\arguments{
\item{t}{A numeric vector representing the input values for \eqn{t \geq 0}.}

\item{mu}{A numeric value representing the input value for \eqn{\mu > 0}.}

\item{phi}{A numeric value representing the input value for \eqn{\phi \geq 0}.}
}
\value{
A numeric vector representing the calculated values of fgp_pg.
}
\description{
Probability generating function of the Geometric Poisson distribution
proposed by Ambagaspitiya and Balakrishnan (1994).
}
\details{
The function \link{fgp_pg} calculates the probability generating by:

\eqn{
 G(t) = e^{-\frac{1}{\phi}\left[W\left(-\frac{\mu\phi}{1 + \mu\phi}t e^{-\frac{\mu\phi}{1 + \mu\phi}}\right) + \frac{\mu\phi}{1 + \mu\phi}\right]},
}
with \eqn{0 \leq t \leq 1}, where \eqn{W} is the Lambert function
(CORLESS et al., 1996), such that, \eqn{W(x)e^{W(x)} = x}.
}
\examples{
fgp_pg(c(0.1, 0.2, 0.3), 2, 0.5)

}
\references{
AMBAGASPITIYA, R. S.; BALAKRISHNAN, N. On the compound generalized
poisson distributions. ASTIN Bulletin: the Journal of the IAA, Cambridge
University Press, v. 24, n. 2, p. 255–263, 1994.

CORLESS, R. M.; GONNET, G. H.; HARE, D. E.; JEFFREY, D. J.; KNUTH,
D. E. On the lambert w function. Advances in Computational mathematics,
Springer, v. 5, p. 329–359, 1996.
}
\seealso{
\link[LambertW:W]{LambertW::W}
}
