% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_function.r
\name{hazard_function}
\alias{hazard_function}
\title{Numerical hazard function}
\usage{
hazard_function(pdf)
}
\arguments{
\item{pdf}{The probability density function used to model the survival time.}
}
\value{
A function that calculates the hazard function at a given time.
}
\description{
This function calculates the hazard function based on a given probability density function (pdf).
The hazard function represents the instantaneous rate at which an event occurs at a specific time,
given that the event has not occurred before that time.
}
\details{
The hazard function is calculated as the ratio of the probability density function
to the survival function. The hazard function is useful for understanding the instantaneous
rate of occurrence of an event at a specific time. The hazard function is also used to model
the instantaneous rate of occurrence of an event in survival analysis. Just like
the function \code{\link[=survival_function]{survival_function()}}, the \code{\link[=hazard_function]{hazard_function()}} function is
vectorized in \eqn{t}, which allows the calculation of the hazard function
for a vector of times. The hazard function is defined by:

\eqn{h(t) = \frac{f_T(t)}{S_T(t)},}
where \eqn{f_T(t)} is the probability density function and \eqn{S_T(t)} is the
survival function at time \eqn{t}.
}
\examples{
# Saraless Nadarajah and Samnuel Kotz (2006)
beta_exponential <- function(x, a, b, lambda){
 lambda / beta(a, b) * exp(-b * lambda * x) * (1 - exp(-lambda * x))^(a - 1)
}

hazard_beta_exponential <- hazard_function(beta_exponential)

hazard_beta_exponential(
  t = seq(0.001, 1.5, length.out = 50L),
  a = 1.5,
  b = 1.8,
  lambda = 1.5
)

}
\references{
NADARAJAH, Saralees; KOTZ, Samuel. The beta exponential distribution. Reliability engineering & system safety, v. 91, n. 6, p. 689-697, 2006.
}
\seealso{
\code{\link[=survival_function]{survival_function()}}.
}
