% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.r
\name{log_likelihood}
\alias{log_likelihood}
\title{Calculate the log-likelihood of a probability density function (pdf).}
\usage{
log_likelihood(pdf)
}
\arguments{
\item{pdf}{A probability density function.}
}
\value{
The log-likelihood function.
}
\description{
This function calculates the log-likelihood of a given probability density function (pdf)
using the negative sum of the logarithm of the pdf evaluated at each value of t.
}
