% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_fragility_zmpg.r
\name{pdf_fragility_zmpg}
\alias{pdf_fragility_zmpg}
\title{Zero-Modified Generalized Poisson Discrete Frailty Model}
\usage{
pdf_fragility_zmpg(pdf)
}
\arguments{
\item{pdf}{The base probability density function.}
}
\value{
Calculates the improper density of a random variable \eqn{T} that
follows the forward discrete frailty model \eqn{f}.
}
\description{
This function constructs the probability density function of the discrete
frailty model of a continuous random variable \eqn{T} through a base density.
}
\details{
The probability density function of a random variable \eqn{T}
requires the base survival function and hazard function, i.e.,
\eqn{S_0(t)} and \eqn{h_0(t)}, respectively. However, it will only be
necessary to inform the base density function as an argument for pdf.
Internally, the function \code{\link[=pdf_fragility_zmpg]{pdf_fragility_zmpg()}} will calculate
\eqn{S_0(t)} and \eqn{h_0(t)}. The probability density function of \eqn{T}
is said to be the discrete frailty model. \eqn{T} is a continuous random
variable, where the density (frailty model) was obtained through a discrete
random variable \eqn{Z} that follows a Zero-Modified Generalized Poisson - ZMPG
distribution, with \eqn{Z \sim ZMPG(\mu, \phi, \rho)}. The probability
density function of the frailty model is given by:

\eqn{

 f_T(t) = -\frac{\rho h_0(t) e^{-\frac{1}{\phi}\left[W\left(-\frac{\mu\phi}{1 + \mu\phi}S_0(t) e^{-\frac{\mu\phi}{1 + \mu\phi}}\right) + \frac{\mu\phi}{1 + \mu\phi}\right]}}{\phi}\frac{W\left(-\frac{\mu\phi}{1 + \mu\phi}S_0(t)e^{-\frac{\mu\phi}{1 + \mu\phi}}\right)}{1 + W\left(-\frac{\mu\phi}{1 + \mu\phi}S_0(t)e^{-\frac{\mu\phi}{1 + \mu\phi}}\right)},

}

with \eqn{t>0}, \eqn{\mu>0}, \eqn{\phi \geq 0} and \eqn{\rho \geq 0}.
}
\examples{
zmpg_weibull <- pdf_fragility_zmpg(pdf = dweibull)
zmpg_weibull(t = 1, mu = 1, phi = 1, rho = 0.5, shape = 0.5, scale = 1.2)
}
\seealso{
\code{\link[=pf_pg]{pf_pg()}}
}
