% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_pg.r
\name{pf_pg}
\alias{pf_pg}
\title{Probability function of a random variable with Generalized Poisson distribution}
\usage{
pf_pg(y, mu, phi)
}
\arguments{
\item{y}{Vector of integer values, with \eqn{y \geq 0}.}

\item{mu}{The mean parameter, with \eqn{\mu > 0}.}

\item{phi}{The dispersion parameter, with \eqn{\phi \geq 0}.}
}
\value{
Probability value of \eqn{Y = y}.
}
\description{
Probability function of a random variable with Generalized Poisson distribution
}
\details{
The probability function of a random variable \eqn{Y} with a Generalized Poisson distribution is expressed by:
\deqn{\pi(y; \mu, \phi) = \frac{\frac{(1 + \phi y)^{y-1}}{y!} \left[\frac{\mu e^{-\mu\phi(1 + \mu\phi)^{-1}}}{1 + \mu\phi}\right]^y}{e^{\mu(1 + \mu\phi)^{-1}}},\,\, y = 0, 1, 2, \ldots,}
\eqn{\mu>0} and \eqn{\phi \geq 0}.
}
\examples{
pf_pg(0:100, 3, 0.5) |> sum()

}
