% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cdf_function.r
\name{plot.cdf_function}
\alias{plot.cdf_function}
\title{Plot Cumulative Distribution Function}
\usage{
\method{plot}{cdf_function}(x, ...)
}
\arguments{
\item{x}{A numeric vector representing the dataset.}

\item{...}{Additional arguments to be passed to the ggplot function.}
}
\value{
A ggplot object displaying the CDF plot.
}
\description{
This function plots the cumulative distribution function (CDF) of a given dataset.
}
\examples{
cdf_weibull <- cdf_function(dweibull)
# plot.cdf_function(cdf_weibull(0:10, shape = 2, scale = 1))

}
