% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hazard_function.r
\name{plot.hazard_function}
\alias{plot.hazard_function}
\title{Plot of the hazard function}
\usage{
\method{plot}{hazard_function}(x, ...)
}
\arguments{
\item{x}{A numeric vector representing the dataset.}

\item{...}{Additional arguments to be passed to the ggplot function.}
}
\value{
A ggplot object displaying the hazard function plot.
}
\description{
Constructs the plot of the hazard function of an object of the
\code{hazard_function} class.
}
\examples{
# Saraless Nadarajah and Samnuel Kotz (2006)
beta_exponential <- function(x, a, b, lambda){
 lambda / beta(a, b) * exp(-b * lambda * x) * (1 - exp(-lambda * x))^(a - 1)
}

hazard_beta_exponential <- hazard_function(beta_exponential)

hazard_beta_exponential(
  t = seq(0.001, 1.5, length.out = 100L),
  a = 1.5,
  b = 1.8,
  lambda = 1.5
) |> plot()

}
\references{
NADARAJAH, Saralees; KOTZ, Samuel. The beta exponential distribution. Reliability engineering & system safety, v. 91, n. 6, p. 689-697, 2006.
}
