% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pdf_fragility_zmpg.r
\name{plot.pdf_fragility_zmpg}
\alias{plot.pdf_fragility_zmpg}
\title{Plot of the probability density function (PDF) of ZMPG frailty.}
\usage{
\method{plot}{pdf_fragility_zmpg}(x, ...)
}
\arguments{
\item{x}{A numeric vector representing the dataset.}

\item{...}{Additional arguments to be passed to the ggplot function.}
}
\value{
A ggplot object displaying the PDF plot.
}
\description{
Constructs a plot of the probability density function (PDF) of ZMPG frailty.
}
\examples{
zmpg_weibull <- pdf_fragility_zmpg(pdf = dweibull)
zmpg_weibull(
  t = seq(0.001, 0.5, length.out = 100L),
  mu = 0.7,
  phi = 0.6,
  rho = 0.1,
  shape = 0.5,
  scale = 1.2
) |> plot()
}
