% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival_function.r
\name{plot.survival_function}
\alias{plot.survival_function}
\title{Plot of the survival function}
\usage{
\method{plot}{survival_function}(x, ...)
}
\arguments{
\item{x}{A numeric vector representing the dataset.}

\item{...}{Additional arguments to be passed to the ggplot function.}
}
\value{
A ggplot object displaying the survival function plot.
}
\description{
Constructs the plot of the survival function of an object of the
\code{survival_function} class.
}
\examples{
survival_weibull <- survival_function(dweibull)
plot.survival_function(survival_weibull(0:10, shape = 2, scale = 1))

}
