% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hazard_function.r
\name{plot.hazard_function}
\alias{plot.hazard_function}
\title{Plot of the hazard function}
\usage{
\method{plot}{hazard_function}(x, ...)
}
\arguments{
\item{x}{A numeric vector representing the dataset.}

\item{...}{Additional arguments to be passed to the ggplot function.}
}
\value{
A ggplot object displaying the hazard function plot.
}
\description{
Constructs the plot of the hazard function of an object of the
\code{hazard_function} class.
}
\examples{
hazard_weibull <- hazard_function(dweibull)
plot.hazard_function(hazard_weibull(1:10, shape = 2, scale = 1))

}
